# codecontext.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CodeContext API",
    "description": "Local code context API for LLMs - provides dependency graph and code relationships",
    "contact": {
      "name": "CodeContext Team",
      "url": "https://github.com/you/codecontext"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:7890",
      "description": "Local CodeContext server"
    }
  ],
  "paths": {
...
    "/api/context/complete": {
      "get": {
        "tags": [
          "codecontext"
        ],
        "summary": "Get complete context for an identifier",
        "description": "Returns comprehensive context information for a code construct. Searches by name or file path. Without 'type' parameter, searches across all entity types and returns best match or multiple matches. With 'type' parameter, searches only within specified type. Returns array of matches to handle ambiguity.",
        "operationId": "GetCompleteContext",
        "parameters": [
          {
            "name": "identifier",
            "in": "query",
            "description": "Name or file path to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by type (Class, Method, Interface, Property, etc.)",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "depth",
            "in": "query",
            "description": "How many relationship levels to traverse (0-10)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "includeTests",
            "in": "query",
            "description": "Whether to include test-related information",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "includeContent",
            "in": "query",
            "description": "Whether to include file content snippets",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "exact",
            "in": "query",
            "description": "Exact matching; omit for exact-first matching with substring fallback",
            "schema": {
              "type": "boolean",
              "default": null
            }
          },
          {
            "name": "includeRelated",
            "in": "query",
            "description": "Whether to include loosely related same-file/namespace symbols",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "includeMetrics",
            "in": "query",
            "description": "Whether to include heuristic code metrics",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "maxMatches",
            "in": "query",
            "description": "Maximum ambiguous candidate summaries to return (1-50)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          },
          {
            "name": "maxRelationships",
            "in": "query",
            "description": "Maximum entries returned per relationship list (1-100)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "maxCallSites",
            "in": "query",
            "description": "Maximum locations returned per aggregated relationship (0-100)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 3
            }
          },
          {
            "name": "expandAmbiguous",
            "in": "query",
            "description": "Expand bounded ambiguous matches instead of returning summaries",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "qualifiedIdentifier",
            "in": "query",
            "description": "Stable qualified identity from an ambiguous summary",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "containingType",
            "in": "query",
            "description": "Filter members by containing type",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "namespace",
            "in": "query",
            "description": "Filter by exact namespace or module",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "signature",
            "in": "query",
            "description": "Filter by exact signature",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "sourceFile",
            "in": "query",
            "description": "Filter by repository-relative or absolute source file",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "view",
            "in": "query",
            "description": "Response shape: compact (default) or full",
            "schema": {
              "type": "string",
              "default": "compact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactContextResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/context/multi": {
      "post": {
        "tags": [
          "codecontext"
        ],
        "summary": "Get context for multiple identifiers",
        "description": "Returns complete context information for multiple code constructs in a single request. Useful for batch operations and reducing HTTP round-trips.",
        "operationId": "GetMultipleContext",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompactContextResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/index/refresh": {
      "post": {
        "tags": [
          "codecontext"
        ],
        "summary": "Refresh the code index",
        "description": "Triggers a re-parse of the specified file or the entire codebase. If path is provided, only that file is refreshed synchronously. If path is omitted, a full background rescan starts (202) and the response carries an operationId observable via /api/status; returns 409 if a scan is already in progress.",
        "operationId": "RefreshIndex",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Path to the file to refresh. If omitted, performs full scan.",
            "schema": {
              "type": "string",
              "default": null
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshFileResponseDto"
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshStartedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/syntax-tree": {
      "post": {
        "tags": [
          "codecontext"
        ],
        "summary": "Get a language-native syntax tree",
        "description": "Returns an on-demand, parser-specific syntax tree for an indexed file. Use the normalized context endpoints for cross-language relationships; use this endpoint for exact syntax, tokens, nesting, and language-specific constructs.",
        "operationId": "GetNativeSyntaxTree",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeSyntaxTreeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NativeSyntaxTreeResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "501": {
            "description": "Not Implemented"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
...
    "/api/shutdown": {
      "post": {
        "tags": [
          "codecontext"
        ],
        "summary": "Gracefully stop this instance",
        "description": "Requests a graceful shutdown of this CodeContext instance. Requires the target instance's instanceId as a query parameter. The instance unregisters itself from the instance registry on exit.",
        "operationId": "Shutdown",
        "parameters": [
          {
            "name": "instanceId",
            "in": "query",
            "schema": {
              "type": "string",
              "default": null
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShutdownResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompactCodeNode": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "file": {
            "type": "string",
            "nullable": true
          },
          "line": {
            "type": "integer",
            "format": "int32"
          },
          "signature": {
            "type": "string",
            "nullable": true
          },
          "relations": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "occurrences": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "callSiteCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "callSitesTruncated": {
            "type": "boolean"
          },
          "distance": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "relationPath": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "qualifiedIdentifier": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CompactContextMatch": {
        "type": "object",
        "properties": {
          "target": {
            "$ref": "#/components/schemas/CompactCodeNode"
          },
          "relationships": {
            "$ref": "#/components/schemas/CompactRelationships"
          },
          "testing": {
            "$ref": "#/components/schemas/CompactTesting"
          },
          "metrics": {
            "$ref": "#/components/schemas/ContextMetrics"
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CompactContextResponse": {
        "type": "object",
        "properties": {
          "view": {
            "type": "string"
          },
          "matchMode": {
            "type": "string"
          },
          "substringSearchSkipped": {
            "type": "boolean"
          },
          "totalMatches": {
            "type": "integer",
            "format": "int32"
          },
          "returnedMatches": {
            "type": "integer",
            "format": "int32"
          },
          "truncated": {
            "type": "boolean"
          },
          "ambiguous": {
            "type": "boolean"
          },
          "disambiguationHint": {
            "type": "string",
            "nullable": true
          },
          "facets": {
            "$ref": "#/components/schemas/CompactMatchFacets"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompactContextMatch"
            }
          }
        }
      },
      "CompactMatchFacets": {
        "type": "object",
        "properties": {
          "types": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "files": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "totalFiles": {
            "type": "integer",
            "format": "int32"
          }
        },
        "nullable": true
      },
...
      "CompactTestFile": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string"
          },
          "testCount": {
            "type": "integer",
            "format": "int32"
          },
          "testMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CodeNode"
            }
          },
          "evidence": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CompactTesting": {
        "type": "object",
        "properties": {
          "isTested": {
            "type": "boolean"
          },
          "directlyTested": {
            "type": "boolean"
          },
          "testReferenceCount": {
            "type": "integer",
            "format": "int32"
          },
          "testImplementerCount": {
            "type": "integer",
            "format": "int32"
          },
          "heuristicMatchCount": {
            "type": "integer",
            "format": "int32"
          },
          "testFileCount": {
            "type": "integer",
            "format": "int32"
          },
          "testFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompactTestFile"
            }
          },
          "truncated": {
            "type": "boolean"
          }
        },
        "nullable": true
      },
      "ContextMetrics": {
        "type": "object",
        "properties": {
          "complexity": {
            "type": "integer",
            "format": "int32"
          },
          "linesOfCode": {
            "type": "integer",
            "format": "int32"
          },
          "dependencyCount": {
            "type": "integer",
            "format": "int32"
          },
          "dependentCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "nullable": true
      },
      "ContextResponseView": {
        "enum": [
          "Compact",
          "Full"
        ]
      },
...
      "JsonElement": { },
      "MultiContextRequest": {
        "type": "object",
        "properties": {
          "identifiers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          },
          "view": {
            "$ref": "#/components/schemas/ContextResponseView"
          },
          "includeTests": {
            "type": "boolean"
          },
          "includeContent": {
            "type": "boolean"
          },
          "exact": {
            "type": "boolean",
            "nullable": true
          },
          "includeRelated": {
            "type": "boolean"
          },
          "includeMetrics": {
            "type": "boolean"
          },
          "maxMatches": {
            "type": "integer",
            "format": "int32"
          },
          "maxRelationships": {
            "type": "integer",
            "format": "int32"
          },
          "maxCallSites": {
            "type": "integer",
            "format": "int32"
          },
          "expandAmbiguous": {
            "type": "boolean"
          },
          "qualifiedIdentifier": {
            "type": "string",
            "nullable": true
          },
          "containingType": {
            "type": "string",
            "nullable": true
          },
          "namespace": {
            "type": "string",
            "nullable": true
          },
          "signature": {
            "type": "string",
            "nullable": true
          },
          "sourceFile": {
            "type": "string",
            "nullable": true
          },
          "relationshipTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "NativeSyntaxTreeRequestDto": {
        "required": [
          "filePath"
        ],
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string"
          },
          "start": {
            "type": "integer",
            "format": "int32",
            "default": null,
            "nullable": true
          },
          "length": {
            "type": "integer",
            "format": "int32",
            "default": null,
            "nullable": true
          },
          "maxDepth": {
            "type": "integer",
            "format": "int32",
            "default": 8
          }
        }
      },
      "NativeSyntaxTreeResult": {
        "required": [
          "parserId",
          "parserVersion",
          "workspaceId",
          "filePath",
          "format",
          "tree",
          "truncated"
        ],
        "type": "object",
        "properties": {
          "parserId": {
            "type": "string"
          },
          "parserVersion": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "tree": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "truncated": {
            "type": "boolean"
          }
        }
      },
...
      "RefreshFileResponseDto": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "RefreshStartedResponseDto": {
        "required": [
          "message",
          "operationId"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "operationId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ShutdownResponseDto": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}